+++ rust/dbn/src/publishers.rs
@@ -155,10 +155,12 @@ pub enum Venue {
     Xtkt = 70,
     /// Osaka Exchange
     Xose = 71,
+    /// Databento Core Indices
+    Dbix = 72,
 }
 
 /// The number of [`Venue`] variants.
-pub const VENUE_COUNT: usize = 71;
+pub const VENUE_COUNT: usize = 72;
 
 impl Venue {
     /// Converts the venue to its `str` representation.
@@ -235,6 +237,7 @@ impl Venue {
             Self::Xtks => "XTKS",
             Self::Xtkt => "XTKT",
             Self::Xose => "XOSE",
+            Self::Dbix => "DBIX",
         }
     }
 }
@@ -327,6 +330,7 @@ impl std::str::FromStr for Venue {
             "XTKS" => Ok(Self::Xtks),
             "XTKT" => Ok(Self::Xtkt),
             "XOSE" => Ok(Self::Xose),
+            "DBIX" => Ok(Self::Dbix),
             _ => Err(Error::conversion::<Self>(s)),
         }
     }
@@ -445,10 +449,12 @@ pub enum Dataset {
     XtktItch = 51,
     /// JPX JGATE3.0-ITCH
     XoseItch = 52,
+    /// Databento Core Indices
+    DbixCore = 53,
 }
 
 /// The number of [`Dataset`] variants.
-pub const DATASET_COUNT: usize = 52;
+pub const DATASET_COUNT: usize = 53;
 
 impl Dataset {
     /// Converts the dataset to its `str` representation.
@@ -508,6 +514,7 @@ impl Dataset {
             Self::XtksFlex => "XTKS.FLEX",
             Self::XtktItch => "XTKT.ITCH",
             Self::XoseItch => "XOSE.ITCH",
+            Self::DbixCore => "DBIX.CORE",
         }
     }
 
@@ -671,6 +678,15 @@ impl Dataset {
             Self::XtksFlex => &[Publisher::XtksFlexXtks],
             Self::XtktItch => &[Publisher::XtktItchXtkt],
             Self::XoseItch => &[Publisher::XoseItchXose],
+            Self::DbixCore => &[
+                Publisher::DbixCoreMain,
+                Publisher::DbixCoreMsci,
+                Publisher::DbixCoreFtse,
+                Publisher::DbixCoreInav,
+                Publisher::DbixCoreMstar,
+                Publisher::DbixCoreCccy,
+                Publisher::DbixCoreCgi,
+            ],
         }
     }
 }
@@ -746,6 +762,7 @@ impl std::str::FromStr for Dataset {
             "XTKS.FLEX" => Ok(Self::XtksFlex),
             "XTKT.ITCH" => Ok(Self::XtktItch),
             "XOSE.ITCH" => Ok(Self::XoseItch),
+            "DBIX.CORE" => Ok(Self::DbixCore),
             _ => Err(Error::conversion::<Self>(s)),
         }
     }
@@ -1048,10 +1065,24 @@ pub enum Publisher {
     XtktItchXtkt = 144,
     /// JPX - Osaka Exchange
     XoseItchXose = 145,
+    /// CGIF - Main Feed
+    DbixCoreMain = 146,
+    /// CGIF - Morgan Stanley Capital International
+    DbixCoreMsci = 147,
+    /// CGIF - FTSE Russell
+    DbixCoreFtse = 148,
+    /// CGIF - Intraday Net Asset Values
+    DbixCoreInav = 149,
+    /// CGIF - Morning Star
+    DbixCoreMstar = 150,
+    /// CGIF - Crypto Currency
+    DbixCoreCccy = 151,
+    /// CGIF - CBOE Global Indices
+    DbixCoreCgi = 152,
 }
 
 /// The number of [`Publisher`] variants.
-pub const PUBLISHER_COUNT: usize = 145;
+pub const PUBLISHER_COUNT: usize = 152;
 
 impl Publisher {
     /// Converts the publisher to its `str` representation.
@@ -1202,6 +1233,13 @@ impl Publisher {
             Self::XtksFlexXtks => "XTKS.FLEX.XTKS",
             Self::XtktItchXtkt => "XTKT.ITCH.XTKT",
             Self::XoseItchXose => "XOSE.ITCH.XOSE",
+            Self::DbixCoreMain => "DBIX.CORE.MAIN",
+            Self::DbixCoreMsci => "DBIX.CORE.MSCI",
+            Self::DbixCoreFtse => "DBIX.CORE.FTSE",
+            Self::DbixCoreInav => "DBIX.CORE.INAV",
+            Self::DbixCoreMstar => "DBIX.CORE.MSTAR",
+            Self::DbixCoreCccy => "DBIX.CORE.CCCY",
+            Self::DbixCoreCgi => "DBIX.CORE.CGI",
         }
     }
 
@@ -1353,6 +1391,13 @@ impl Publisher {
             Self::XtksFlexXtks => Venue::Xtks,
             Self::XtktItchXtkt => Venue::Xtkt,
             Self::XoseItchXose => Venue::Xose,
+            Self::DbixCoreMain => Venue::Main,
+            Self::DbixCoreMsci => Venue::Msci,
+            Self::DbixCoreFtse => Venue::Ftse,
+            Self::DbixCoreInav => Venue::Inav,
+            Self::DbixCoreMstar => Venue::Mstar,
+            Self::DbixCoreCccy => Venue::Cccy,
+            Self::DbixCoreCgi => Venue::Cgi,
         }
     }
 
@@ -1504,6 +1549,13 @@ impl Publisher {
             Self::XtksFlexXtks => Dataset::XtksFlex,
             Self::XtktItchXtkt => Dataset::XtktItch,
             Self::XoseItchXose => Dataset::XoseItch,
+            Self::DbixCoreMain => Dataset::DbixCore,
+            Self::DbixCoreMsci => Dataset::DbixCore,
+            Self::DbixCoreFtse => Dataset::DbixCore,
+            Self::DbixCoreInav => Dataset::DbixCore,
+            Self::DbixCoreMstar => Dataset::DbixCore,
+            Self::DbixCoreCccy => Dataset::DbixCore,
+            Self::DbixCoreCgi => Dataset::DbixCore,
         }
     }
 
@@ -1658,6 +1710,13 @@ impl Publisher {
             (Dataset::XtksFlex, Venue::Xtks) => Ok(Self::XtksFlexXtks),
             (Dataset::XtktItch, Venue::Xtkt) => Ok(Self::XtktItchXtkt),
             (Dataset::XoseItch, Venue::Xose) => Ok(Self::XoseItchXose),
+            (Dataset::DbixCore, Venue::Main) => Ok(Self::DbixCoreMain),
+            (Dataset::DbixCore, Venue::Msci) => Ok(Self::DbixCoreMsci),
+            (Dataset::DbixCore, Venue::Ftse) => Ok(Self::DbixCoreFtse),
+            (Dataset::DbixCore, Venue::Inav) => Ok(Self::DbixCoreInav),
+            (Dataset::DbixCore, Venue::Mstar) => Ok(Self::DbixCoreMstar),
+            (Dataset::DbixCore, Venue::Cccy) => Ok(Self::DbixCoreCccy),
+            (Dataset::DbixCore, Venue::Cgi) => Ok(Self::DbixCoreCgi),
             _ => Err(Error::conversion::<Self>(format_args!(
                 "({dataset}, {venue})"
             ))),
@@ -1827,6 +1886,13 @@ impl std::str::FromStr for Publisher {
             "XTKS.FLEX.XTKS" => Ok(Self::XtksFlexXtks),
             "XTKT.ITCH.XTKT" => Ok(Self::XtktItchXtkt),
             "XOSE.ITCH.XOSE" => Ok(Self::XoseItchXose),
+            "DBIX.CORE.MAIN" => Ok(Self::DbixCoreMain),
+            "DBIX.CORE.MSCI" => Ok(Self::DbixCoreMsci),
+            "DBIX.CORE.FTSE" => Ok(Self::DbixCoreFtse),
+            "DBIX.CORE.INAV" => Ok(Self::DbixCoreInav),
+            "DBIX.CORE.MSTAR" => Ok(Self::DbixCoreMstar),
+            "DBIX.CORE.CCCY" => Ok(Self::DbixCoreCccy),
+            "DBIX.CORE.CGI" => Ok(Self::DbixCoreCgi),
             _ => Err(Error::conversion::<Self>(s)),
         }
     }